+++ components/org.wso2.carbon.identity.conditional.auth.functions.session/src/test/java/org/wso2/carbon/identity/conditional/auth/functions/session/test/SessionModelTest.java
@@ -16,11 +16,11 @@
  * under the License.
  *
  */
-package org.wso2.carbon.identity.extensions.conditional.auth.functions.test;
+package org.wso2.carbon.identity.conditional.auth.functions.session.test;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.wso2.carbon.identity.extensions.authenticator.conditional.auth.functions.model.Session;
+import org.wso2.carbon.identity.conditional.auth.functions.session.model.Session;
 
 /**
  * Contains methods for testing Session Model class methods.
